Skip tests removed from Module::CPANTS::Analyse
Some of the metrics will be removed (or moved to Module::CPANTS::SiteKwalitee) in the near future when the next version of Module::CPANTS::Analyse is released. This patch checks which tests are done and skips the rest in the end.

Or, you might want to declare the number of tests that actually remain in Module::CPANTS::Analyse/Kwalitee. Whichever is ok, as long as the number of the tests is correctly handled.